rag_sql_app/requirements.txt
openai
chromadb
sentence-transformers
torch
numpy<2.0
psycopg2-binary